backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
cachetools>=5.3.0
supabase>=2.0.0
yfinance>=0.2.31
pandas>=2.0.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
